+++ inc/sync_queue.h
@@ -18,10 +18,10 @@ struct sync_queue
 	pthread_cond_t q_full_wait;
 
 	// this is the number of threads waiting on the q_empty_wait
-	unsigned int q_empty_wait_thread_count;
+	uint64_t q_empty_wait_thread_count;
 
 	// this is the number of threads waiting on the q_full_wait
-	unsigned int q_full_wait_thread_count;
+	uint64_t q_full_wait_thread_count;
 
 	// this value suggests the maximum capacity of this queue
 	cy_uint max_capacity;
@@ -110,8 +110,8 @@ void close_sync_queue(sync_queue* sq);
 /*
 *	TO CHECK THE NUMBER OF CONSUMERS AND PRODUCERS WAITING ON THE CONDITIONAL WAITS OF SYNC_QUEUE
 */
-unsigned int get_threads_waiting_on_empty_sync_queue(sync_queue* sq);
+uint64_t get_threads_waiting_on_empty_sync_queue(sync_queue* sq);
 
-unsigned int get_threads_waiting_on_full_sync_queue(sync_queue* sq);
+uint64_t get_threads_waiting_on_full_sync_queue(sync_queue* sq);
 
 #endif
